Add LicenseField
This doesn't yet properly update the record, but does show the options
correctly.

I've added a search prop to the DropDown as there are so many options in the
vocabulary.

It doesn't use SelectField, as there's not the same mapping between what is
selected, and what needs to be inserted into the record.

Partially resolves #11.